+++ lib/charliecloud.py
@@ -1,38 +1,24 @@
-import abc
 import argparse
 import atexit
 import collections
 import collections.abc
-import copy
 import cProfile
 import datetime
 import enum
-import errno
-import fcntl
-import getpass
 import hashlib
 import io
-import json
 import os
-import pathlib
 import platform
-import pprint
 import pstats
 import re
 import shlex
 import shutil
 import signal
-import stat
-import struct
 import subprocess
 import sys
-import tarfile
 import time
 import traceback
 import types
-import urllib
-
-import version
 
 
 ## Hairy imports ##
@@ -114,27 +100,6 @@ ARCH_MAP = { "x86_64":    "amd64",
 ARCH_MAP_FALLBACK = { "arm/v7": ("arm",),
                       "arm64/v8": ("arm64",) }
 
-# ARGs that are “magic”: always available, don’t cause cache misses, not saved
-# with the image.
-ARGS_MAGIC = { "HTTP_PROXY", "HTTPS_PROXY", "FTP_PROXY", "NO_PROXY",
-               "http_proxy", "https_proxy", "ftp_proxy", "no_proxy",
-               "SSH_AUTH_SOCK", "USER" }
-# FIXME: user() not yet defined
-ARG_DEFAULTS_MAGIC = { k:v for (k,v) in ((m, os.environ.get(m))
-                                          for m in ARGS_MAGIC)
-                       if v is not None }
-
-# ARGs with pre-defined default values that *are* saved with the image.
-ARG_DEFAULTS = \
-   { # calls to chown/fchown withn a user namespace will fail with EINVAL for
-     # UID/GIDs besides the current one. This env var tells fakeroot to not
-     # try. Credit to Dave Dykstra for pointing us to this.
-     "FAKEROOTDONTTRYCHOWN": "1",
-     "PATH": "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-     # GNU tar, when it thinks it’s running as root, tries to chown(2) and
-     # chgrp(2) files to whatever is in the tarball.
-     "TAR_OPTIONS": "--no-same-owner" }
-
 # String to use as hint when we throw an error that suggests a bug.
 BUG_REPORT_PLZ = "please report this bug: https://github.com/hpc/charliecloud/issues"
 
@@ -145,147 +110,8 @@ BUG_REPORT_PLZ = "please report this bug: https://github.com/hpc/charliecloud/is
 # somewhat random guess with hopefully enough headroom not to cause problems.
 FILENAME_MAX_CHARS = 192
 
-# Common grammar stuff.
-GRAMMAR_COMMON = r"""
-// Matching lines in the face of continuations is surprisingly hairy. Notes:
-//
-//   1. The underscore prefix means the rule is always inlined (i.e., removed
-//      and children become children of its parent).
-//
-//   2. LINE_CHUNK must not match any characters that _LINE_CONTINUE does.
-//
-//   3. This is very sensitive to the location of repetition. Moving the plus
-//      either to the entire regex (i.e., “/(...)+/”) or outside the regex
-//      (i.e., ”/.../+”) gave parse errors.
-//
-_line: ( _LINE_CONTINUE | LINE_CHUNK )+
-LINE_CHUNK: /[^\\\n]+|(\\(?![ \t]+\n))+/
-
-HEX_STRING: /[0-9A-Fa-f]+/
-WORD: /[^ \t\n=]/+
-
-IR_PATH_COMPONENT: /[a-z0-9_.-]+/
-
-_string_list: "[" _WS? STRING_QUOTED ( "," _WS? STRING_QUOTED )* _WS? "]"
-
-_WSH: /[ \t]/+                   // sequence of horizontal whitespace
-_LINE_CONTINUE: "\\" _WSH? "\n"  // line continuation
-_WS: ( _WSH | _LINE_CONTINUE )+  // horizontal whitespace w/ line continuations
-_NEWLINES: ( _WS? "\n" )+        // sequence of newlines
-
-%import common.ESCAPED_STRING -> STRING_QUOTED
-"""
-
-# Dockerfile grammar. Note image references are not parsed during Dockerfile
-# parsing.
-GRAMMAR_DOCKERFILE = r"""
-start: dockerfile
-
-// First instruction must be ARG or FROM, but that is not a syntax error.
-dockerfile: _NEWLINES? ( arg_first | directive | comment )* ( instruction | comment )*
-
-?instruction: _WS? ( arg | copy | env | from_ | run | shell | workdir | uns_forever | uns_yet )
-
-directive.2: _WS? "#" _WS? DIRECTIVE_NAME "=" _line _NEWLINES
-DIRECTIVE_NAME: ( "escape" | "syntax" )
-
-comment: _WS? _COMMENT_BODY _NEWLINES
-_COMMENT_BODY: /#[^\n]*/
-
-arg: "ARG"i _WS ( arg_bare | arg_equals ) _NEWLINES
-arg_bare: WORD
-arg_equals: WORD "=" ( WORD | STRING_QUOTED )
-
-arg_first.2: "ARG"i _WS ( arg_first_bare | arg_first_equals ) _NEWLINES
-arg_first_bare: WORD
-arg_first_equals: WORD "=" ( WORD | STRING_QUOTED )
-
-copy: "COPY"i ( _WS option )* _WS ( copy_list | copy_shell ) _NEWLINES
-copy_list.2: _string_list
-copy_shell: WORD ( _WS WORD )+
-
-env: "ENV"i _WS ( env_space | env_equalses ) _NEWLINES
-env_space: WORD _WS _line
-env_equalses: env_equals ( _WS env_equals )*
-env_equals: WORD "=" ( WORD | STRING_QUOTED )
-
-from_: "FROM"i ( _WS ( option | option_keypair ) )* _WS image_ref [ _WS from_alias ] _NEWLINES
-from_alias: "AS"i _WS IR_PATH_COMPONENT  // FIXME: undocumented; this is guess
-
-run: "RUN"i _WS ( run_exec | run_shell ) _NEWLINES
-run_exec.2: _string_list
-run_shell: _line
-
-shell: "SHELL"i _WS _string_list _NEWLINES
-
-workdir: "WORKDIR"i _WS _line _NEWLINES
-
-uns_forever: UNS_FOREVER _WS _line _NEWLINES
-UNS_FOREVER: ( "EXPOSE"i | "HEALTHCHECK"i | "MAINTAINER"i | "STOPSIGNAL"i | "USER"i | "VOLUME"i )
-
-uns_yet: UNS_YET _WS _line _NEWLINES
-UNS_YET: ( "ADD"i | "CMD"i | "ENTRYPOINT"i | "LABEL"i | "ONBUILD"i )
-
-/// Common ///
-
-option: "--" OPTION_KEY "=" OPTION_VALUE
-option_keypair: "--" OPTION_KEY "=" OPTION_VAR "=" OPTION_VALUE
-OPTION_KEY: /[a-z]+/
-OPTION_VALUE: /[^= \t\n]+/
-OPTION_VAR: /[a-z]+/
-
-image_ref: IMAGE_REF
-IMAGE_REF: /[A-Za-z0-9$:._\/-]+/
-""" + GRAMMAR_COMMON
-
-# Grammar for image references.
-GRAMMAR_IMAGE_REF = r"""
-// Note: Hostnames with no dot and no port get parsed as a hostname, which
-// is wrong; it should be the first path component. We patch this error later.
-// FIXME: Supposedly this can be fixed with priorities, but I couldn't get it
-// to work with brief trying.
-
-start: image_ref
-
-image_ref: ir_hostport? ir_path? ir_name ( ir_tag | ir_digest )?
-ir_hostport: IR_HOST ( ":" IR_PORT )? "/"
-ir_path: ( IR_PATH_COMPONENT "/" )+
-ir_name: IR_PATH_COMPONENT
-ir_tag: ":" IR_TAG
-ir_digest: "@sha256:" HEX_STRING
-IR_HOST: /[A-Za-z0-9_.-]+/
-IR_PORT: /[0-9]+/
-IR_TAG: /[A-Za-z0-9_.-]+/
-""" + GRAMMAR_COMMON
-
-# Chunk size in bytes when streaming HTTP. Progress meter is updated once per
-# chunk, which means the display is updated roughly every 20s at 100 Kbit/s
-# and every 2s at 1Mbit/s; beyond that, the once-per-second display throttling
-# takes over.
-HTTP_CHUNK_SIZE = 256 * 1024
-
 # Minimum Python version. NOTE: Keep in sync with configure.ac.
 PYTHON_MIN = (3,6)
-
-# Content types for some stuff we care about.
-# See: https://github.com/opencontainers/image-spec/blob/main/media-types.md
-TYPES_MANIFEST = \
-   {"docker2": "application/vnd.docker.distribution.manifest.v2+json",
-    "oci1":    "application/vnd.oci.image.manifest.v1+json"}
-TYPES_INDEX = \
-   {"docker2": "application/vnd.docker.distribution.manifest.list.v2+json",
-    "oci1":    "application/vnd.oci.image.index.v1+json"}
-TYPE_CONFIG = "application/vnd.docker.container.image.v1+json"
-TYPE_LAYER = "application/vnd.docker.image.rootfs.diff.tar.gzip"
-
-# Top-level directories we create if not present.
-STANDARD_DIRS = { "bin", "dev", "etc", "mnt", "proc", "sys", "tmp", "usr" }
-
-# Storage directory format version. We refuse to operate on storage
-# directories with non-matching versions. Increment this number when the
-# format changes non-trivially.
-STORAGE_VERSION = 4
-
 
 ## Globals ##
 
@@ -303,23 +129,12 @@ log_festoon = False  # If true, prepend pid and timestamp to chatter.
 log_fp = sys.stderr  # File object to print logs to.
 trace_fatal = False  # Add abbreviated traceback to fatal error hint.
 
-# Verify TLS certificates? Passed to requests.
-tls_verify = True
-
 # True if the download cache is enabled.
 dlcache_p = None
 
 # Profiling.
 profiling = False
 profile = None
-
-# True if we talk to registries authenticated; false if anonymously.
-reg_auth = False
-
-# True if we lock storage directory to prevent concurrent access; false for no
-# locking (which is very YOLO and may break the storage directory).
-storage_lock = True
-
 
 ## Exceptions ##
 
@@ -411,725 +226,6 @@ class ArgumentParser(argparse.ArgumentParser):
       return cli
 
 
-class Credentials:
-
-   __slots__ = ("password",
-                "username")
-
-   def __init__(self):
-      self.username = None
-      self.password = None
-
-   def get(self):
-      # If stored, return those.
-      if (self.username is not None):
-         username = self.username
-         password = self.password
-      else:
-         try:
-            # Otherwise, use environment variables.
-            username = os.environ["CH_IMAGE_USERNAME"]
-            password = os.environ["CH_IMAGE_PASSWORD"]
-         except KeyError:
-            # Finally, prompt the user.
-            # FIXME: This hangs in Bats despite sys.stdin.isatty() == True.
-            try:
-               username = input("\nUsername: ")
-            except KeyboardInterrupt:
-               FATAL("authentication cancelled")
-            password = getpass.getpass("Password: ")
-         if (not password_many):
-            # Remember the credentials.
-            self.username = username
-            self.password = password
-      return (username, password)
-
-
-class Image:
-   """Container image object.
-
-      Constructor arguments:
-
-        ref........... Image_Ref object to identify the image.
-
-        unpack_path .. Directory to unpack the image in; if None, infer path
-                       in storage dir from ref."""
-
-   __slots__ = ("metadata",
-                "ref",
-                "unpack_path")
-
-   def __init__(self, ref, unpack_path=None):
-      assert isinstance(ref, Image_Ref)
-      self.ref = ref
-      if (unpack_path is not None):
-         assert isinstance(unpack_path, Path)
-         self.unpack_path = unpack_path
-      else:
-         self.unpack_path = storage.unpack(self.ref)
-      self.metadata_init()
-
-   @property
-   def deleteable(self):
-      """True if it's OK to delete me, either my unpack directory (a) is at
-         the expected location within the storage directory xor (b) is not not
-         but it looks like an image; False otherwise."""
-      if (self.unpack_path == storage.unpack_base // self.unpack_path.name):
-         return True
-      else:
-         if (all(os.path.isdir(self.unpack_path // i)
-                for i in ("bin", "dev", "usr"))):
-            return True
-      return False
-
-   @property
-   def last_modified(self):
-      # Return the last modified time of self as a datetime.datetime object in
-      # the local time zone.
-      return datetime.datetime.fromtimestamp(
-                 (self.metadata_path // "metadata.json").stat_(False).st_mtime,
-                 datetime.timezone.utc).astimezone()
-
-   @property
-   def metadata_path(self):
-      return self.unpack_path // "ch"
-
-   @property
-   def unpack_exist_p(self):
-      return os.path.exists(self.unpack_path)
-
-   @property
-   def unpack_cache_linked(self):
-      return self.unpack_exist_p and os.path.exists(self.unpack_path // ".git")
-
-   def __str__(self):
-      return str(self.ref)
-
-   @classmethod
-   def glob(class_, image_glob):
-      """Return a possibly-empty iterator of images in the storage directory
-         matching the given glob."""
-      for ref in Image_Ref.glob(image_glob):
-         yield class_(ref)
-
-   def commit(self):
-      "Commit the current unpack directory into the layer cache."
-      assert False, "unimplemented"
-
-   def copy_unpacked(self, other):
-      """Copy image other to my unpack directory, which may not exist. other
-         can be either a path (string or Path object) or an Image object; in
-         the latter case other.unpack_path is used. other need not be a valid
-         image; the essentials will be created if needed."""
-      def ignore(path, names):
-         # Need this path conversion, since shutil._copytree() calls ignore()
-         # and passes a string, not a Path. See:
-         # https://github.com/python/cpython/blob/main/Lib/shutil.py#L456
-         path = Path(path)
-         ignore = list()
-         if (path == src_path):
-            for name in names:
-               if (name.startswith(".git")):
-                  VERBOSE("ignoring from source image: /%s" % name)
-                  ignore.append(name)
-         return ignore
-      if (isinstance(other, str) or isinstance(other, Path)):
-         src_path = other
-      else:
-         src_path = other.unpack_path
-      VERBOSE("copying image: %s -> %s" % (src_path, self.unpack_path))
-      Path(src_path).copytree(self.unpack_path, symlinks=True, ignore=ignore)
-      self.unpack_init()
-
-   def layers_open(self, layer_tars):
-      """Open the layer tarballs and read some metadata (which unfortunately
-         means reading the entirety of every file). Return an OrderedDict:
-
-           keys:    layer hash (full)
-           values:  namedtuple with two fields:
-                      fp:       open TarFile object
-                      members:  sequence of members (OrderedSet)
-
-         Empty layers are skipped.
-
-         Important note: TarFile.extractall() extracts the given members in
-         the order they are specified, so we need to preserve their order from
-         the file, as returned by getmembers(). We also need to quickly remove
-         members we don't want from this sequence. Thus, we use the OrderedSet
-         class defined in this module."""
-      TT = collections.namedtuple("TT", ["fp", "members"])
-      layers = collections.OrderedDict()
-      # Schema version one (v1) allows one or more empty layers for Dockerfile
-      # entries like CMD (https://github.com/containers/skopeo/issues/393).
-      # Unpacking an empty layer doesn't accomplish anything so we ignore them.
-      empty_cnt = 0
-      for (i, path) in enumerate(layer_tars, start=1):
-         lh = os.path.basename(path).split(".", 1)[0]
-         lh_short = lh[:7]
-         INFO("layer %d/%d: %s: listing" % (i, len(layer_tars), lh_short))
-         try:
-            fp = TarFile.open(path)
-            members = OrderedSet(fp.getmembers())  # reads whole file :(
-         except tarfile.TarError as x:
-            FATAL("cannot open: %s: %s" % (path, x))
-         if (lh in layers and len(members) > 0):
-            FATAL("duplicate non-empty layer %s" % lh)
-         if (len(members) > 0):
-            layers[lh] = TT(fp, members)
-         else:
-            empty_cnt += 1
-      VERBOSE("skipped %d empty layers" % empty_cnt)
-      return layers
-
-   def metadata_init(self):
-      "Initialize empty metadata structure."
-      # Elsewhere can assume the existence and types of everything here.
-      self.metadata = { "arch": arch_host.split("/")[0],  # no variant
-                        "arg": { **ARG_DEFAULTS_MAGIC, **ARG_DEFAULTS },
-                        "cwd": "/",
-                        "env": dict(),
-                        "history": list(),
-                        "labels": dict(),
-                        "shell": ["/bin/sh", "-c"],
-                        "volumes": list() }  # set isn't JSON-serializable
-
-   def metadata_load(self, target_img=None):
-      """Load metadata file, replacing the existing metadata object. If
-         metadata doesn't exist, warn and use defaults. If target_img is
-         non-None, use that image's metadata instead of self's."""
-      if (target_img is not None):
-         path = target_img.metadata_path
-      else:
-         path = self.metadata_path
-      path //= "metadata.json"
-      if (path.exists()):
-         VERBOSE("loading metadata")
-      else:
-         WARNING("no metadata to load; using defaults")
-         self.metadata_init()
-         return
-      self.metadata = path.json_from_file("metadata")
-      # upgrade old metadata
-      self.metadata.setdefault("arg", dict())
-      self.metadata.setdefault("history", list())
-      # add default ARG variables
-      self.metadata["arg"].update({ **ARG_DEFAULTS_MAGIC, **ARG_DEFAULTS })
-
-   def metadata_merge_from_config(self, config):
-      """Interpret all the crap in the config data structure that is meaingful
-         to us, and add it to self.metadata. Ignore anything we expect in
-         config that's missing."""
-      def get(*keys):
-         d = config
-         keys = list(keys)
-         while (len(keys) > 1):
-            try:
-               d = d[keys.pop(0)]
-            except KeyError:
-               return None
-         assert (len(keys) == 1)
-         return d.get(keys[0])
-      def set_(dst_key, *src_keys):
-         v = get(*src_keys)
-         if (v is not None and v != ""):
-            self.metadata[dst_key] = v
-      if ("config" not in config):
-         FATAL("config missing key 'config'")
-      # architecture
-      set_("arch", "architecture")
-      # $CWD
-      set_("cwd", "config", "WorkingDir")
-      # environment
-      env = get("config", "Env")
-      if (env is not None):
-         for line in env:
-            try:
-               (k,v) = line.split("=", maxsplit=1)
-            except AttributeError:
-               FATAL("can't parse config: bad Env line: %s" % line)
-            self.metadata["env"][k] = v
-      # History.
-      if ("history" not in config):
-         FATAL("invalid config: missing history")
-      self.metadata["history"] = config["history"]
-      # labels
-      set_("labels", "config", "Labels")  # copy reference
-      # shell
-      set_("shell", "config", "Shell")
-      # Volumes. FIXME: Why is this a dict with empty dicts as values?
-      vols = get("config", "Volumes")
-      if (vols is not None):
-         for k in config["config"]["Volumes"].keys():
-            self.metadata["volumes"].append(k)
-
-   def metadata_replace(self, config_json):
-      self.metadata_init()
-      if (config_json is None):
-         INFO("no config found; initializing empty metadata")
-      else:
-         # Copy pulled config file into the image so we still have it.
-         path = self.metadata_path // "config.pulled.json"
-         copy2(config_json, path)
-         VERBOSE("pulled config path: %s" % path)
-         self.metadata_merge_from_config(path.json_from_file("config"))
-      self.metadata_save()
-
-   def metadata_save(self):
-      """Dump image's metadata to disk, including the main data structure but
-         also all auxiliary files, e.g. ch/environment."""
-      # Adjust since we don't save everything.
-      metadata = copy.deepcopy(self.metadata)
-      for k in ARGS_MAGIC:
-         metadata["arg"].pop(k, None)
-      # Serialize. We take care to pretty-print this so it can (sometimes) be
-      # parsed by simple things like grep and sed.
-      out = json.dumps(metadata, indent=2, sort_keys=True)
-      DEBUG("metadata:\n%s" % out)
-      # Main metadata file.
-      path = self.metadata_path // "metadata.json"
-      VERBOSE("writing metadata file: %s" % path)
-      path.file_write(out + "\n")
-      # /ch/environment
-      path = self.metadata_path // "environment"
-      VERBOSE("writing environment file: %s" % path)
-      path.file_write( (  "\n".join("%s=%s" % (k,v) for (k,v)
-                                    in sorted(metadata["env"].items()))
-                        + "\n"))
-      # mkdir volumes
-      VERBOSE("ensuring volume directories exist")
-      for path in metadata["volumes"]:
-         (self.unpack_path // path).mkdirs()
-
-   def tarballs_write(self, tarball_dir):
-      """Write one uncompressed tarball per layer to tarball_dir. Return a
-         sequence of tarball basenames, with the lowest layer first."""
-      # FIXME: Yes, there is only one layer for now and we'll need to update
-      # it when (if) we have multiple layers. But, I wanted the interface to
-      # support multiple layers.
-      base = "%s.tar" % self.ref.for_path
-      path = tarball_dir // base
-      try:
-         INFO("layer 1/1: gathering")
-         VERBOSE("writing tarball: %s" % path)
-         fp = TarFile.open(path, "w", format=tarfile.PAX_FORMAT)
-         unpack_path = self.unpack_path.resolve()  # aliases use symlinks
-         VERBOSE("canonicalized unpack path: %s" % unpack_path)
-         fp.add_(unpack_path, arcname=".")
-         fp.close()
-      except OSError as x:
-         FATAL("can't write tarball: %s" % x.strerror)
-      return [base]
-
-   def unpack(self, layer_tars, last_layer=None):
-      """Unpack config_json (path to JSON config file) and layer_tars
-         (sequence of paths to tarballs, with lowest layer first) into the
-         unpack directory, validating layer contents and dealing with
-         whiteouts. Empty layers are ignored. The unpack directory must exist,
-         and either be empty or contain only a single entry called “.git”."""
-      if (last_layer is None):
-         last_layer = sys.maxsize
-      INFO("flattening image")
-      self.unpack_layers(layer_tars, last_layer)
-      self.unpack_init()
-
-   def unpack_cache_unlink(self):
-      (self.unpack_path // ".git").unlink()
-
-   def unpack_clear(self):
-      """If the unpack directory does not exist, do nothing. If the unpack
-         directory is already an image, remove it. Otherwise, error."""
-      if (not os.path.exists(self.unpack_path)):
-         VERBOSE("no image found: %s" % self.unpack_path)
-      else:
-         if (not os.path.isdir(self.unpack_path)):
-            FATAL("can't flatten: %s exists but is not a directory"
-                  % self.unpack_path)
-         if (not self.deleteable):
-            FATAL("can't flatten: %s exists but does not appear to be an image"
-                  % self.unpack_path)
-         VERBOSE("removing image: %s" % self.unpack_path)
-         t = Timer()
-         self.unpack_path.rmtree()
-         t.log("removed image")
-
-   def unpack_delete(self):
-      VERBOSE("unpack path: %s" % self.unpack_path)
-      if (not self.unpack_exist_p):
-         FATAL("image not found, can’t delete: %s" % self.ref)
-      if (self.deleteable):
-         INFO("deleting image: %s" % self.ref)
-         self.unpack_path.chmod_min(0o700)
-         for (dir_, subdirs, _) in os.walk(self.unpack_path):
-            # must fix as subdirs so we can traverse into them
-            for subdir in subdirs:
-               (Path(dir_) // subdir).chmod_min(0o700)
-         self.unpack_path.rmtree()
-      else:
-         FATAL("storage directory seems broken: not an image: %s" % self.ref)
-
-   def unpack_init(self):
-      """Initialize the unpack directory, which must exist. Any setup already
-         present will be left unchanged. After this, self.unpack_path is a
-         valid Charliecloud image directory."""
-      # Metadata directory.
-      (self.unpack_path // "ch").mkdir_()
-      (self.unpack_path // "ch/environment").file_ensure_exists()
-      # Essential directories & mount points. Do nothing if something already
-      # exists, without dereferencing, in case it's a symlink, which will work
-      # for bind-mount later but won't resolve correctly now outside the
-      # container (e.g. linuxcontainers.org images; issue #1015).
-      #
-      # WARNING: Keep in sync with shell scripts.
-      for d in list(STANDARD_DIRS) + ["mnt/%d" % i for i in range(10)]:
-         d = self.unpack_path // d
-         if (not os.path.lexists(d)):
-            d.mkdirs()
-      (self.unpack_path // "etc/hosts").file_ensure_exists()
-      (self.unpack_path // "etc/resolv.conf").file_ensure_exists()
-
-   def unpack_layers(self, layer_tars, last_layer):
-      layers = self.layers_open(layer_tars)
-      self.validate_members(layers)
-      self.whiteouts_resolve(layers)
-      self.unpack_path.mkdir_()  # create directory in case no layers
-      for (i, (lh, (fp, members))) in enumerate(layers.items(), start=1):
-         lh_short = lh[:7]
-         if (i > last_layer):
-            INFO("layer %d/%d: %s: skipping per --last-layer"
-                 % (i, len(layers), lh_short))
-         else:
-            INFO("layer %d/%d: %s: extracting" % (i, len(layers), lh_short))
-            try:
-               fp.extractall(path=self.unpack_path, members=members)
-            except OSError as x:
-               FATAL("can't extract layer %d: %s" % (i, x.strerror))
-
-   def validate_members(self, layers):
-      INFO("validating tarball members")
-      top_dirs = set()
-      VERBOSE("pass 1: canonicalizing member paths")
-      for (i, (lh, (fp, members))) in enumerate(layers.items(), start=1):
-         abs_ct = 0
-         for m in list(members):   # copy b/c we remove items from the set
-            # Remove members with empty paths.
-            if (len(m.name) == 0):
-               WARNING("layer %d/%d: %s: skipping member with empty path"
-                       % (i, len(layers), lh[:7]))
-               members.remove(m)
-            # Convert member paths to Path objects for easier processing.
-            # Note: In my testing, parsing a string into a Path object took
-            # about 2.5µs, so this should be plenty fast.
-            m.name = Path(m.name)
-            # Reject members with up-levels.
-            if (".." in m.name.parts):
-               FATAL("rejecting up-level member: %s: %s" % (fp.name, m.name))
-            # Correct absolute paths.
-            if (m.name.is_absolute()):
-               m.name = m.name.relative_to("/")
-            # Record top-level directory.
-            if (len(m.name.parts) > 1 or m.isdir()):
-               top_dirs.add(m.name.first)
-         if (abs_ct > 0):
-            WARNING("layer %d/%d: %s: fixed %d absolute member paths"
-                    % (i, len(layers), lh[:7], abs_ct))
-      top_dirs.discard(None)  # ignore “.”
-      # Convert to tarbomb if (1) there is a single enclosing directory and
-      # (2) that directory is not one of the standard directories, e.g. to
-      # allow images containing just “/bin/fooprog”.
-      if (len(top_dirs) != 1 or not top_dirs.isdisjoint(STANDARD_DIRS)):
-         VERBOSE("pass 2: conversion to tarbomb not needed")
-      else:
-         VERBOSE("pass 2: converting to tarbomb")
-         for (i, (lh, (fp, members))) in enumerate(layers.items(), start=1):
-            for m in members:
-               if (len(m.name.parts) > 0):  # ignore “.”
-                  m.name = Path(*m.name.parts[1:])  # strip first component
-      VERBOSE("pass 3: analyzing members")
-      for (i, (lh, (fp, members))) in enumerate(layers.items(), start=1):
-         dev_ct = 0
-         link_fix_ct = 0
-         for m in list(members):  # copy again
-            m.name = str(m.name)  # other code assumes strings
-            if (m.isdev()):
-               # Device or FIFO: Ignore.
-               dev_ct += 1
-               VERBOSE("ignoring device file: %s" % m.name)
-               members.remove(m)
-               continue
-            elif (m.issym() or m.islnk()):
-               link_fix_ct += TarFile.fix_link_target(m, fp.name)
-            elif (m.isdir()):
-               # Directory: Fix bad permissions (hello, Red Hat).
-               m.mode |= 0o700
-            elif (m.isfile()):
-               # Regular file: Fix bad permissions (HELLO RED HAT!!).
-               m.mode |= 0o600
-            else:
-               FATAL("unknown member type: %s" % m.name)
-            # Discard Git metadata (files that begin with “.git”).
-            if (re.search(r"^(\./)?\.git", m.name)):
-               WARNING("ignoring member: %s" % m.name)
-               members.remove(m)
-               continue
-            # Discard anything under /dev. Docker puts regular files and
-            # directories in here on "docker export". Note leading slashes
-            # already taken care of in TarFile.fix_member_path() above.
-            if (re.search(r"^(\./)?dev/.", m.name)):
-               VERBOSE("ignoring member under /dev: %s" % m.name)
-               members.remove(m)
-               continue
-            TarFile.fix_member_uidgid(m)
-         if (dev_ct > 0):
-            WARNING("layer %d/%d: %s: ignored %d devices and/or FIFOs"
-                    % (i, len(layers), lh[:7], dev_ct))
-         if (link_fix_ct > 0):
-            INFO("layer %d/%d: %s: changed %d absolute symbolic and/or hard links to relative"
-                 % (i, len(layers), lh[:7], link_fix_ct))
-
-   def whiteout_rm_prefix(self, layers, max_i, prefix):
-      """Ignore members of all layers from 1 to max_i inclusive that have path
-         prefix of prefix. For example, if prefix is foo/bar, then ignore
-         foo/bar and foo/bar/baz but not foo/barbaz. Return count of members
-         ignored."""
-      TRACE("finding members with prefix: %s" % prefix)
-      prefix = os.path.normpath(prefix)  # "./foo" == "foo"
-      ignore_ct = 0
-      for (i, (lh, (fp, members))) in enumerate(layers.items(), start=1):
-         if (i > max_i): break
-         members2 = list(members)  # copy b/c we'll alter members
-         for m in members2:
-            if (prefix_path(prefix, m.name)):
-               ignore_ct += 1
-               members.remove(m)
-               TRACE("layer %d/%d: %s: ignoring %s"
-                     % (i, len(layers), lh[:7], m.name))
-      return ignore_ct
-
-   def whiteouts_resolve(self, layers):
-      """Resolve whiteouts. See:
-         https://github.com/opencontainers/image-spec/blob/master/layer.md"""
-      INFO("resolving whiteouts")
-      for (i, (lh, (fp, members))) in enumerate(layers.items(), start=1):
-         wo_ct = 0
-         ig_ct = 0
-         members2 = list(members)  # copy b/c we'll alter members
-         for m in members2:
-            dir_ = os.path.dirname(m.name)
-            filename = os.path.basename(m.name)
-            if (filename.startswith(".wh.")):
-               wo_ct += 1
-               members.remove(m)
-               if (filename == ".wh..wh..opq"):
-                  # "Opaque whiteout": remove contents of dir_.
-                  DEBUG("found opaque whiteout: %s" % m.name)
-                  ig_ct += self.whiteout_rm_prefix(layers, i - 1, dir_)
-               else:
-                  # "Explicit whiteout": remove same-name file without ".wh.".
-                  DEBUG("found explicit whiteout: %s" % m.name)
-                  ig_ct += self.whiteout_rm_prefix(layers, i - 1,
-                                                   dir_ + "/" + filename[4:])
-         if (wo_ct > 0):
-            VERBOSE("layer %d/%d: %s: %d whiteouts; %d members ignored"
-                    % (i, len(layers), lh[:7], wo_ct, ig_ct))
-
-
-class Image_Ref:
-   """Reference to an image in a remote repository.
-
-      The constructor takes one argument, which is interpreted differently
-      depending on type:
-
-        None or omitted... Build an empty Image_Ref (all fields None).
-
-        string ........... Parse it; see FAQ for syntax. Can be either the
-                           standard form (e.g., as in a FROM instruction) or
-                           our filename form with percents replacing slashes.
-
-        Lark parse tree .. Must be same result as parsing a string. This
-                           allows the parse step to be embedded in a larger
-                           parse (e.g., a Dockerfile).
-
-     Warning: References containing a hostname without a dot and no port
-     cannot be round-tripped through a string, because the hostname will be
-     assumed to be a path component."""
-
-   __slots__ = ("host",
-                "port",
-                "path",
-                "name",
-                "tag",
-                "digest",
-                "variables")
-
-   # Reference parser object. Instantiating a parser took 100ms when we tested
-   # it, which means we can't really put it in a loop. But, at parse time,
-   # "lark" may refer to a dummy module (see above), so we can't populate the
-   # parser here either. We use a class varible and populate it at the time of
-   # first use.
-   parser = None
-
-   def __init__(self, src=None, variables=None):
-      self.host = None
-      self.port = None
-      self.path = []
-      self.name = None
-      self.tag = None
-      self.digest = None
-      self.variables = dict() if variables is None else variables
-      if (isinstance(src, str)):
-         src = self.parse(src, self.variables)
-      if (isinstance(src, lark.tree.Tree)):
-         self.from_tree(src)
-      elif (src is not None):
-         assert False, "unsupported initialization type"
-
-   def __str__(self):
-      out = ""
-      if (self.host is not None):
-         out += self.host
-      if (self.port is not None):
-         out += ":" + str(self.port)
-      if (self.host is not None):
-         out += "/"
-      out += self.path_full
-      if (self.tag is not None):
-         out += ":" + self.tag
-      if (self.digest is not None):
-         out += "@sha256:" + self.digest
-      return out
-
-   @staticmethod
-   def path_to_ref(path):
-      if (isinstance(path, Path)):
-         path = path.name
-      return path.replace("+", ":").replace("%", "/")
-
-   @staticmethod
-   def ref_to_pathstr(ref_str):
-      return ref_str.replace("/", "%").replace(":", "+")
-
-   @classmethod
-   def glob(class_, image_glob):
-      """Return a possibly-empty iterator of references in the storage
-         directory matching the given glob."""
-      for path in storage.unpack_base.glob(class_.ref_to_pathstr(image_glob)):
-         yield class_(class_.path_to_ref(path))
-
-   @classmethod
-   def parse(class_, s, variables):
-      if (class_.parser is None):
-         class_.parser = lark.Lark(GRAMMAR_IMAGE_REF, parser="earley",
-                                   propagate_positions=True)
-      s = s.replace("%", "/").replace("+", ":")
-      hint="https://hpc.github.io/charliecloud/faq.html#how-do-i-specify-an-image-reference"
-      s = variables_sub(s, variables)
-      if "$" in s:
-         FATAL("image reference contains an undefined variable: %s" % s)
-      try:
-         tree = class_.parser.parse(s)
-      except lark.exceptions.UnexpectedInput as x:
-         if (x.column == -1):
-            FATAL("image ref syntax, at end: %s" % s, hint);
-         else:
-            FATAL("image ref syntax, char %d: %s" % (x.column, s), hint)
-      except lark.exceptions.UnexpectedEOF as x:
-         # We get UnexpectedEOF because of Lark issue #237. This exception
-         # doesn't have a column location.
-         FATAL("image ref syntax, at end: %s" % s, hint)
-      DEBUG(tree.pretty())
-      return tree
-
-   @property
-   def as_verbose_str(self):
-      def fmt(x):
-         if (x is None):
-            return None
-         else:
-            return repr(x)
-      return """\
-as string:    %s
-for filename: %s
-fields:
-  host    %s
-  port    %s
-  path    %s
-  name    %s
-  tag     %s
-  digest  %s\
-""" % tuple(  [str(self), self.for_path]
-            + [fmt(i) for i in (self.host, self.port, self.path,
-                                self.name, self.tag, self.digest)])
-
-   @property
-   def canonical(self):
-      "Copy of self with all the defaults filled in."
-      ref = self.copy()
-      ref.defaults_add()
-      return ref
-
-   @property
-   def for_path(self):
-      return self.ref_to_pathstr(str(self))
-
-   @property
-   def path_full(self):
-      out = ""
-      if (len(self.path) > 0):
-         out += "/".join(self.path) + "/"
-      out += self.name
-      return out
-
-   @property
-   def version(self):
-      if (self.tag is not None):
-         return self.tag
-      if (self.digest is not None):
-         return "sha256:" + self.digest
-      assert False, "version invalid with no tag or digest"
-
-   def copy(self):
-      "Return an independent copy of myself."
-      return copy.deepcopy(self)
-
-   def defaults_add(self):
-      "Set defaults for all empty fields."
-      if (self.host is None):
-         if ("CH_REGY_DEFAULT_HOST" not in os.environ):
-            self.host = "registry-1.docker.io"
-         else:
-            self.host = os.getenv("CH_REGY_DEFAULT_HOST")
-            self.port = int(os.getenv("CH_REGY_DEFAULT_PORT", 443))
-            prefix = os.getenv("CH_REGY_PATH_PREFIX")
-            if (prefix is not None):
-               self.path = prefix.split("/") + self.path
-      if (self.port is None): self.port = 443
-      if (self.host == "registry-1.docker.io" and len(self.path) == 0):
-         # FIXME: For Docker Hub only, images with no path need a path of
-         # "library" substituted. Need to understand/document the rules here.
-         self.path = ["library"]
-      if (self.tag is None and self.digest is None): self.tag = "latest"
-
-   def from_tree(self, t):
-      self.host = tree_child_terminal(t, "ir_hostport", "IR_HOST")
-      self.port = tree_child_terminal(t, "ir_hostport", "IR_PORT")
-      if (self.port is not None):
-         self.port = int(self.port)
-      self.path = [    variables_sub(s, self.variables)
-                   for s in tree_child_terminals(t, "ir_path",
-                                                 "IR_PATH_COMPONENT")]
-      self.name = tree_child_terminal(t, "ir_name", "IR_PATH_COMPONENT")
-      self.tag = tree_child_terminal(t, "ir_tag", "IR_TAG")
-      self.digest = tree_child_terminal(t, "ir_digest", "HEX_STRING")
-      for a in ("host", "port", "name", "tag", "digest"):
-         setattr(self, a, variables_sub(getattr(self, a), self.variables))
-      # Resolve grammar ambiguity for hostnames w/o dot or port.
-      if (    self.host is not None
-          and "." not in self.host
-          and self.port is None):
-         self.path.insert(0, self.host)
-         self.host = None
-
-
 class OrderedSet(collections.abc.MutableSet):
 
    # Note: The superclass provides basic implementations of all the other
@@ -1164,396 +260,6 @@ class OrderedSet(collections.abc.MutableSet):
    def discard(self, x):
       self.data.pop(x, None)
 
-
-class Path(pathlib.PosixPath):
-   """Stock Path objects have the very weird property that appending an
-      *absolute* path to an existing path ignores the left operand, leaving
-      only the absolute right operand:
-
-        >>> import pathlib
-        >>> a = pathlib.Path("/foo/bar")
-        >>> a.joinpath("baz")
-        PosixPath('/foo/bar/baz')
-        >>> a.joinpath("/baz")
-        PosixPath('/baz')
-
-      This is contrary to long-standing UNIX/POSIX, where extra slashes in a
-      path are ignored, e.g. the path "foo//bar" is equivalent to "foo/bar".
-      It seems to be inherited from os.path.join().
-
-      Even with the relatively limited use of Path objects so far, this has
-      caused quite a few bugs. IMO it's too difficult and error-prone to
-      manually manage whether paths are absolute or relative. Thus, this
-      subclass introduces a new operator "//" which does the right thing,
-      i.e., if the right operand is absolute, that fact is ignored. E.g.:
-
-        >>> a = Path("/foo/bar")
-        >>> a.joinpath_posix("baz")
-        Path('/foo/bar/baz')
-        >>> a.joinpath_posix("/baz")
-        Path('/foo/bar/baz')
-        >>> a // "/baz"
-        Path('/foo/bar/baz')
-        >>> "/baz" // a
-        Path('/baz/foo/bar')
-
-      We introduce a new operator because it seemed like too subtle a change
-      to the existing operator "/" (which we disable to avoid getting burned
-      here in Charliecloud). An alternative was "+" like strings, but that led
-      to silently wrong results when the paths *were* strings (components
-      concatenated with no slash)."""
-
-   # Name of the gzip(1) to use; set on first call of file_gzip().
-   gzip = None
-
-   def __floordiv__(self, right):
-      return self.joinpath_posix(right)
-
-   def __len__(self):
-      return self.parts.__len__()
-
-   def __rfloordiv__(self, left):
-      left = Path(left)
-      return left.joinpath_posix(self)
-
-   def __rtruediv__(self, left):
-      return NotImplemented
-
-   def __truediv__(self, right):
-      return NotImplemented
-
-   @property
-   def first(self):
-      """Return my first component, or if I have no components (i.e.,
-         Path(".")), return None."""
-      try:
-         return self.parts[0]
-      except IndexError:
-         return None
-
-   @property
-   def git_escaped(self):
-      "Return a copy of me escaped for Git storage."
-      assert (self.git_incompatible_p)
-      return self.with_name(self.name.replace(".git", ".weirdal_"))
-
-   @property
-   def git_incompatible_p(self):
-      "Return True if I can’t be stored in Git because of my name."
-      return self.name.startswith(".git")
-
-   @classmethod
-   def gzip_set(cls):
-      """Set gzip class attribute on first call to file_gzip().
-
-         Note: We originally thought this could be accomplished WITHOUT
-         calling a class method (by setting the attribute, e.g. “self.gzip =
-         'foo'”), but it turned out that this would only set the attribute for
-         the single instance. To set self.gzip for all instances, we need the
-         class method."""
-      if (cls.gzip is None):
-         if (shutil.which("pigz") is not None):
-            cls.gzip = "pigz"
-         elif (shutil.which("gzip") is not None):
-            cls.gzip = "gzip"
-         else:
-            FATAL("can’t find path to gzip or pigz")
-
-   def add_suffix(self, suff):
-      """Returns the path object restulting from appending the specified
-         suffix to the end of the path name. E.g. Path(foo).add_suffix(".txt")
-         returns Path("foo.txt)."""
-      return Path(str(self) + suff)
-
-   def chdir(self):
-      "Change CWD to path and return previous CWD. Exit on error."
-      old = ossafe(os.getcwd, "can’t get cwd(2)")
-      ossafe(os.chdir, "can’t chdir: %s" % self.name, self)
-      return Path(old)
-
-   def chmod_min(self, perms_new, st=None):
-      """Set permissions on path so they are at least mode. If given, st is a
-         stat object for self, to avoid another stat(2) call if unneeded.
-         Return the new file mode (complete, not just permission bits).
-
-         For symlinks, do nothing, because we don’t want to follow symlinks
-         and follow_symlinks=False (or os.lchmod) is not supported on some
-         (all?) Linux. (Also, symlink permissions are ignored on Linux, so it
-         doesn’t matter anyway.)"""
-      if (st is None):
-         st = self.stat_(False)
-      if (stat.S_ISLNK(st.st_mode)):
-         return st.st_mode
-      perms_old = stat.S_IMODE(st.st_mode)
-      perms_new |= perms_old
-      if (perms_new != perms_old):
-         VERBOSE("fixing permissions: %s: %03o -> %03o"
-                 % (self, perms_old, perms_new))
-         ossafe(os.chmod, "can’t chmod: %s" % self, self, perms_new)
-      return (st.st_mode | perms_new)
-
-   def copytree(self, *args, **kwargs):
-      "Wrapper for shutil.copytree() that exits on the first error."
-      shutil.copytree(str(self), copy_function=copy2, *args, **kwargs)
-
-   def disk_bytes(self):
-      """Return the number of disk bytes consumed by path. Note this is probably
-         different from the file size."""
-      return self.stat().st_blocks * 512
-
-   def du(self):
-      """Return a tuple (number of files, total bytes on disk) for everything
-         under path. Warning: double-counts files with multiple hard links."""
-      file_ct = 1
-      byte_ct = self.disk_bytes()
-      for (dir_, subdirs, files) in os.walk(self):
-         file_ct += len(subdirs) + len(files)
-         byte_ct += sum(Path(dir_ + "/" + i).disk_bytes()
-                        for i in subdirs + files)
-      return (file_ct, byte_ct)
-
-   def exists_(self, links=False):
-      "Return True if I exist, False otherwise. Iff links, follow symlinks."
-      try:
-         # Don’t wrap self.exists() because that always follows symlnks. Use
-         # os.stat() b/c it has follow_symlinks in 3.6, unlike self.stat().
-         os.stat(self, follow_symlinks=links)
-      except FileNotFoundError:
-         return False
-      except OSError as x:
-         FATAL("can’t stat: %s: %s" % (self, x.strerror))
-      return True
-
-   def file_ensure_exists(self):
-      """If the final element of path exists (without dereferencing if it’s a
-         symlink), do nothing; otherwise, create it as an empty regular file."""
-      if (not os.path.lexists(self)): # no substitute for lexists() in pathlib.
-         fp = self.open_("w")
-         close_(fp)
-
-   def file_gzip(self, args=[]):
-      """Run pigz(1) if it’s available, otherwise gzip(1), on file at path and
-         return the file's new name. Pass args to the gzip executable. This
-         lets us gzip files (a) in parallel if pigz(1) is installed and
-         (b) without reading them into memory."""
-      path_c = self.add_suffix(".gz")
-      # On first call, remember first available of pigz and gzip using class
-      # attribute 'gzip'.
-      Path.gzip_set()
-      # Remove destination if it already exists, because “gzip --force” does
-      # several other things too. Also, pigz(1) sometimes confusingly reports
-      # “Inappropriate ioctl for device” if destination already exists.
-      if (path_c.exists()):
-         path_c.unlink()
-      # Compress.
-      cmd([self.gzip] + args + [str(self)])
-      # Zero out GZIP header timestamp, bytes 4–7 zero-indexed inclusive [1], to
-      # ensure layer hash is consistent. See issue #1080.
-      # [1]: https://datatracker.ietf.org/doc/html/rfc1952 §2.3.1
-      fp = path_c.open_("r+b")
-      ossafe(fp.seek, "can’t seek: %s" % fp, 4)
-      ossafe(fp.write, "can’t write: %s" % fp, b'\x00\x00\x00\x00')
-      close_(fp)
-      return path_c
-
-   def file_hash(self):
-      """Return the hash of data in file at path, as a hex string with no
-         algorithm tag. File is read in chunks and can be larger than memory."""
-      fp = self.open_("rb")
-      h = hashlib.sha256()
-      while True:
-         data = ossafe(fp.read, "can’t read: %s" % self.name, 2**18)
-         if (len(data) == 0):
-            break  # EOF
-         h.update(data)
-      close_(fp)
-      return h.hexdigest()
-
-   def file_read_all(self, text=True):
-      """Return the contents of file at path, or exit with error. If text, read
-         in "rt" mode with UTF-8 encoding; otherwise, read in mode "rb"."""
-      if (text):
-         mode = "rt"
-         encoding = "UTF-8"
-      else:
-         mode = "rb"
-         encoding = None
-      fp = self.open_(mode, encoding=encoding)
-      data = ossafe(fp.read, "can't read: %s" % self.name)
-      close_(fp)
-      return data
-
-   def file_size(self, follow_symlinks=False):
-      "Return the size of file at path in bytes."
-      st = ossafe(os.stat, "can’t stat: %s" % self.name,
-                  self, follow_symlinks=follow_symlinks)
-      return st.st_size
-
-   def file_write(self, content):
-      if (isinstance(content, str)):
-         content = content.encode("UTF-8")
-      fp = self.open_("wb")
-      ossafe(fp.write, "can’t write: %s" % self.name, content)
-      close_(fp)
-
-   def grep_p(self, rx):
-      """Return True if file at path contains a line matching regular
-         expression rx, False if it does not."""
-      rx = re.compile(rx)
-      try:
-         with open(self, "rt") as fp:
-            for line in fp:
-               if (rx.search(line) is not None):
-                  return True
-         return False
-      except OSError as x:
-         FATAL("can’t read %s: %s" % (self.name, x.strerror))
-
-   def hardlink(self, target):
-      try:
-         os.link(target, self)  # no super().hardlink_to() until 3.10
-      except OSError as x:
-         FATAL("can’t hard link: %s -> %s: %s" % (self, target, x.strerror))
-
-   def is_relative_to(self, *other):
-      try:
-         return super().is_relative_to(*other)
-      except AttributeError:
-         # pathlib.Path.is_relative_to() was introduced in 3.9. If not
-         # available, use the standard library’s trivial definition in terms
-         # of relative_to().
-         try:
-            self.relative_to(*other)
-            return True
-         except ValueError:
-            return False
-
-   def joinpath_posix(self, other):
-      # This method is a hot spot, so the hairiness is due to optimizations.
-      # It runs about 30% faster than the naïve verson below.
-      if (isinstance(other, Path)):
-         other_parts = other._parts
-         if (len(other_parts) > 0 and other_parts[0] == "/"):
-            other_parts = other_parts[1:]
-      elif (isinstance(other, str)):
-         other_parts = other.split("/")
-         if (len(other_parts) > 0 and len(other_parts[0]) == 0):
-            other_parts = other_parts[1:]
-      else:
-         assert False, "unknown type"
-      return self._from_parsed_parts(self._drv, self._root,
-                                     self._parts + other_parts)
-      # Naïve implementation for reference.
-      #other = Path(other)
-      #if (other.is_absolute()):
-      #   other = other.relative_to("/")
-      #   assert (not other.is_absolute())
-      #return self.joinpath(other)
-
-   def json_from_file(self, msg):
-      DEBUG("loading JSON: %s: %s" % (msg, self))
-      text = self.file_read_all()
-      TRACE("text:\n%s" % text)
-      try:
-         data = json.loads(text)
-         DEBUG("result:\n%s" % pprint.pformat(data, indent=2))
-      except json.JSONDecodeError as x:
-         FATAL("can’t parse JSON: %s:%d: %s" % (self.name, x.lineno, x.msg))
-      return data
-
-   def listdir(self):
-      """Return set of entries in directory path, as strings, without self (.)
-         and parent (..). We considered changing this to use os.scandir() for
-         #992, but decided that the advantages it offered didn’t warrant the
-         effort required to make the change."""
-      return set(ossafe(os.listdir, "can’t list: %s" % self.name, self))
-
-   def strip(self, left=0, right=0):
-      """Return a copy of myself with n leading components removed. E.g.:
-
-           >>> a = Path("/a/b/c")
-           >>> a.strip(left=1)
-           Path("a/b/c")
-           >>> a.strip(right=1)
-           Path("/a/b")
-           >>> a.strip(left=1, right=1)
-           Path("a/b")
-
-         It is an error if I don’t have at least left + right components,
-         i.e., you can strip a path down to nothing but not further."""
-      assert (len(self.parts) >= left + right)
-      return Path(*self.parts[left:len(self.parts)-right])
-
-   def mkdir_(self):
-      TRACE("ensuring directory: %s" % self)
-      try:
-         super().mkdir(exist_ok=True)
-      except FileExistsError as x:
-         FATAL("can’t mkdir: exists and not a directory: %s" % x.filename)
-      except OSError as x:
-         FATAL("can’t mkdir: %s: %s: %s" % (self.name, x.filename, x.strerror))
-
-   def mkdirs(self, exist_ok=True):
-      TRACE("ensuring directories: %s" % self.name)
-      try:
-         os.makedirs(self, exist_ok=exist_ok)
-      except OSError as x:
-         FATAL("can’t mkdir: %s: %s: %s" % (self.name, x.filename, x.strerror))
-
-   def open_(self, mode, *args, **kwargs):
-      return ossafe(super().open, "can't open for %s: %s" % (mode, self.name),
-                    mode, *args, **kwargs)
-
-   def rename_(self, name_new):
-      if (Path(name_new).exists()):
-         FATAL("can’t rename: destination exists: %s" % name_new)
-      ossafe(super().rename, "can’t rename: %s -> %s" % (self.name, name_new),
-             name_new)
-
-   def rmdir_(self):
-      ossafe(super().rmdir, "can’t rmdir: %s" % self.name)
-
-   def rmtree(self):
-      if (self.is_dir()):
-         TRACE("deleting directory: %s" % self.name)
-         try:
-            shutil.rmtree(self)
-         except OSError as x:
-            FATAL("can’t recursively delete directory %s: %s: %s"
-                  % (self.name, x.filename, x.strerror))
-      else:
-         assert False, "unimplemented"
-
-   def stat_(self, links):
-      """An error-checking version of stat(). Note that we cannot simply
-         change the definition of stat() to be ossafe, as the exists() method
-         in pathlib relies on an OSError check.
-
-         See: https://github.com/python/cpython/blob/3.10/Lib/pathlib.py#L1291
-
-         NOTE: We also cannot just call super().stat here because the
-         follow_symlinks kwarg is absent in pathlib for Python 3.6, which we
-         want to retain compatibility with."""
-      return ossafe(os.stat, "can’t stat: %s" % self, self,
-                    follow_symlinks=links)
-
-   def symlink(self, target, clobber=False):
-      if (clobber and self.is_file()):
-         self.unlink_()
-      try:
-         super().symlink_to(target)
-      except FileExistsError:
-         if (not self.is_symlink()):
-            FATAL(  "can’t symlink: source exists and isn't a symlink: %s"
-                  % self.name)
-         if (self.readlink() != target):
-            FATAL("can’t symlink: %s exists; want target %s but existing is %s"
-                  % (self.name, target, self.readlink()))
-      except OSError as x:
-         FATAL("can’t symlink: %s -> %s: %s" % (self.name, target, x.strerror))
-
-   def unlink_(self, *args, **kwargs):
-      ossafe(super().unlink, "can't unlink: %s" % self.name)
 
 class Progress:
    """Simple progress meter for countable things that updates at most once per
@@ -1692,867 +398,6 @@ class Progress_Writer:
       self.progress.update(len(data))
       ossafe(self.fp.write, "can't write: %s" % self.path, data)
 
-
-class Registry_Auth(requests.auth.AuthBase):
-
-   # Every registry request has an “authorization object”. This starts as no
-   # authentication at all. If we get HTTP 401 Unauthorized, we try to
-   # “escalate” to a higher level of authorization; some classes have multiple
-   # escalators that we try in order. Escalation can fail either if
-   # authentication fails or there is nothing to escalate to.
-   #
-   # Class attributes:
-   #
-   #   anon_p ...... True if the authorization object is anonymous; False if
-   #                 it needed authentication.
-   #
-   #   escalators .. Sequence of classes we can escalate to. Empty if no
-   #                 escalation possible. This is actually a property rather
-   #                 than a class attribute because it needs to refer to
-   #                 classes that may not have been defined when the module is
-   #                 created, e.g. classes later in the file, or some can
-   #                 escalate to themselves.
-   #
-   #   reg_auth .... True if appropriate for authenticated mode, False if
-   #                 anonymous (i.e., --auth or not, respectively). Everything
-   #                 must be one or the other.
-   #
-   #   scheme ...... Auth scheme string (from WWW-Authenticate header) that
-   #                 this class matches.
-
-   __slots__ = ("auth_h_next",)  # WWW-Authenticate header for next escalator
-
-   def __eq__(self, other):
-      return (type(self) == type(other))
-
-   @property
-   def escalators(self):
-      ...
-
-   @classmethod
-   def authenticate(class_, creds, auth_d):
-      """Authenticate using the given credentials and parsed WWW-Authenticate
-         dictionary. Return a new Registry_Auth object if successful, None if
-         not. The caller is responsible for dealing with the failure."""
-      ...
-
-   def escalate(self, reg, res):
-      """Escalate to a higher level of authorization. Use the WWW-Authenticate
-         header in failed response res if there is one."""
-      VERBOSE("escalating from %s" % self)
-      assert (res.status_code == 401)
-      # Get authentication instructions.
-      if ("WWW-Authenticate" in res.headers):
-         auth_h = res.headers["WWW-Authenticate"]
-      elif (self.auth_h_next is not None):
-         auth_h = self.auth_h_next
-      else:
-         FATAL("don’t know how to authenticate: WWW-Authenticate not found")
-      # We use two “undocumented (although very stable and frequently cited)”
-      # methods to parse the authentication response header (thanks Andy,
-      # i.e., @adrecord on GitHub).
-      (auth_scheme, auth_d) = auth_h.split(maxsplit=1)
-      auth_d = urllib.request.parse_keqv_list(
-                  urllib.request.parse_http_list(auth_d))
-      VERBOSE("WWW-Authenticate parsed: %s %s" % (auth_scheme, auth_d))
-      # Is escalation possible in principle?
-      if (len(self.escalators) == 0):
-         FATAL("no further authentication possible, giving up")
-      # Try to escalate.
-      for class_ in self.escalators:
-         if (class_.scheme == auth_scheme):
-            if (class_.reg_auth != reg_auth):
-               VERBOSE("skipping %s: auth mode mismatch" % class_.__name__)
-            else:
-               VERBOSE("authenticating using %s" % class_.__name__)
-               auth = class_.authenticate(reg, auth_d)
-               if (auth is None):
-                  VERBOSE("authentication failed; trying next")
-               elif (auth == self):
-                  VERBOSE("authentication did not escalate; trying next")
-               else:
-                  return auth  # success!
-      VERBOSE("no authentication left to try")
-      return None
-
-
-class Registry_Auth_Basic(Registry_Auth):
-   anon_p = False
-   scheme = "Basic"
-   reg_auth = True
-
-   __slots__ = ("basic")
-
-   def __call__(self, *args, **kwargs):
-      return self.basic(*args, **kwargs)
-
-   def __eq__(self, other):
-      return super().__eq__(other) and (self.basic == other.basic)
-
-   def __str__(self):
-      return self.basic.__str__()
-
-   @property
-   def escalators(self):
-      return ()
-
-   @classmethod
-   def authenticate(class_, reg, auth_d):
-      # Note: Basic does not validate the credentials until we try to use it.
-      if ("realm" not in auth_d):
-         FATAL("WWW-Authenticate missing realm")
-      (username, password) = reg.creds.get()
-      i = class_()
-      i.basic = requests.auth.HTTPBasicAuth(username, password)
-      return i
-
-
-class Registry_Auth_Bearer_IDed(Registry_Auth):
-   # https://stackoverflow.com/a/58055668
-   anon_p = False
-   scheme = "Bearer"
-   reg_auth = True
-   variant = "IDed"
-
-   __slots__ = ("auth_d",
-                "token")
-
-   def __init__(self, token, auth_d):
-      self.token = token
-      self.auth_d = auth_d
-
-   def __call__(self, req):
-      req.headers["Authorization"] = "Bearer %s" % self.token
-      return req
-
-   def __eq__(self, other):
-      return super().__eq__(other) and (self.auth_d == other.auth_d)
-
-   def __str__(self):
-      return ("Bearer (%s) %s" % (self.__class__.__name__.split("_")[-1],
-                                  self.token_short))
-
-   @property
-   def escalators(self):
-      # One can escalate to an authenticated Bearer with a greater scope. I’m
-      # pretty sure this doesn’t create an infinite loop because eventually
-      # the token request will fail.
-      return (Registry_Auth_Bearer_IDed,)
-
-   @property
-   def token_short(self):
-      return ("%s..%s" % (self.token[:8], self.token[-8:]))
-
-   @classmethod
-   def authenticate(class_, reg, auth_d):
-      # Registries and endpoints vary in what they put in WWW-Authenticate. We
-      # need realm because it's the URL to use for a token. Otherwise, just
-      # give back all the keys we got.
-      for k in ("realm",):
-         if (k not in auth_d):
-            FATAL("WWW-Authenticate missing key: %s" % k)
-      params = { (k,v) for (k,v) in auth_d.items() if k != "realm" }
-      # Request a Bearer token.
-      res = reg.request_raw("GET", auth_d["realm"], {200,401,403},
-                            auth=class_.token_auth(reg.creds), params=params)
-      if (res.status_code != 200):
-         VERBOSE("bearer token request rejected")
-         return None
-      # Create new instance.
-      i = class_(res.json()["token"], auth_d)
-      VERBOSE("received bearer token: %s" % (i.token_short))
-      return i
-
-   @classmethod
-   def token_auth(class_, creds):
-      """Return a requests.auth.AuthBase object used to authenticate the token
-         request."""
-      (username, password) = creds.get()
-      return requests.auth.HTTPBasicAuth(username, password)
-
-
-class Registry_Auth_Bearer_Anon(Registry_Auth_Bearer_IDed):
-   anon_p = True
-   scheme = "Bearer"
-   reg_auth = False
-
-   __slots__ = ()
-
-   @property
-   def escalators(self):
-      return (Registry_Auth_Bearer_IDed,)
-
-   @classmethod
-   def token_auth(class_, creds):
-      # The way to get an anonymous Bearer token is to give no Basic auth
-      # header in the token request.
-      return None
-
-
-class Registry_Auth_None(Registry_Auth):
-   anon_p = True
-   scheme = None
-   #reg_auth =   # not meaningful b/c we start here in both modes
-
-   def __call__(self, req):
-      return req
-
-   def __str__(self):
-      return "no authorization"
-
-   @property
-   def escalators(self):
-      return (Registry_Auth_Basic,
-              Registry_Auth_Bearer_Anon,
-              Registry_Auth_Bearer_IDed)
-
-
-class Registry_HTTP:
-   """Transfers image data to and from a remote image repository via HTTPS.
-
-      Note that ref refers to the *remote* image. Objects of this class have
-      no information about the local image."""
-
-   __slots__ = ("auth",
-                "creds",
-                "ref",
-                "session")
-
-   def __init__(self, ref):
-      # Need an image ref with all the defaults filled in.
-      self.ref = ref.canonical
-      self.auth = Registry_Auth_None()
-      self.creds = Credentials()
-      self.session = None
-      # This is commented out because it prints full request and response
-      # bodies to standard output (not stderr), which overwhelms the terminal.
-      # Normally, a better debugging approach if you need this is to sniff the
-      # connection using e.g. mitmproxy.
-      #if (verbose >= 2):
-      #   http.client.HTTPConnection.debuglevel = 1
-
-   def _url_of(self, type_, address):
-      "Return an appropriate repository URL."
-      url_base = "https://%s:%d/v2" % (self.ref.host, self.ref.port)
-      return "/".join((url_base, self.ref.path_full, type_, address))
-
-   def blob_exists_p(self, digest):
-      """Return true if a blob with digest (hex string) exists in the
-         remote repository, false otherwise."""
-      # Gotchas:
-      #
-      # 1. HTTP 401 means both unauthorized *or* not found, I assume to avoid
-      #    information leakage about the presence of stuff one isn't allowed
-      #    to see. By the time it gets here, we should be authenticated, so
-      #    interpret it as not found.
-      #
-      # 2. Sometimes we get 301 Moved Permanently. It doesn't bubble up to
-      #    here because requests.request() follows redirects. However,
-      #    requests.head() does not follow redirects, and it seems like a
-      #    weird status, so I worry there is a gotcha I haven't figured out.
-      url = self._url_of("blobs", "sha256:%s" % digest)
-      res = self.request("HEAD", url, {200,401,404})
-      return (res.status_code == 200)
-
-   def blob_to_file(self, digest, path, msg):
-      "GET the blob with hash digest and save it at path."
-      # /v2/library/hello-world/blobs/<layer-hash>
-      url = self._url_of("blobs", "sha256:" + digest)
-      sw = Progress_Writer(path, msg)
-      self.request("GET", url, out=sw)
-      sw.close()
-
-   def blob_upload(self, digest, data, note=""):
-      """Upload blob with hash digest to url. data is the data to upload, and
-         can be anything requests can handle; if it's an open file, then it's
-         wrapped in a Progress_Reader object. note is a string to prepend to
-         the log messages; default empty string."""
-      INFO("%s%s: checking if already in repository" % (note, digest[:7]))
-      # 1. Check if blob already exists. If so, stop.
-      if (self.blob_exists_p(digest)):
-         INFO("%s%s: already present" % (note, digest[:7]))
-         return
-      msg = "%s%s: not present, uploading" % (note, digest[:7])
-      if (isinstance(data, io.IOBase)):
-         data = Progress_Reader(data, msg)
-         data.start()
-      else:
-         INFO(msg)
-      # 2. Get upload URL for blob.
-      url = self._url_of("blobs", "uploads/")
-      res = self.request("POST", url, {202})
-      # 3. Upload blob. We do a "monolithic" upload (i.e., send all the
-      # content in a single PUT request) as opposed to a "chunked" upload
-      # (i.e., send data in multiple PATCH requests followed by a PUT request
-      # with no body).
-      url = res.headers["Location"]
-      res = self.request("PUT", url, {201}, data=data,
-                         params={ "digest": "sha256:%s" % digest })
-      if (isinstance(data, Progress_Reader)):
-         data.close()
-      # 4. Verify blob now exists.
-      if (not self.blob_exists_p(digest)):
-         FATAL("blob just uploaded does not exist: %s" % digest[:7])
-
-   def close(self):
-      if (self.session is not None):
-         self.session.close()
-
-   def config_upload(self, config):
-      "Upload config (sequence of bytes)."
-      self.blob_upload(bytes_hash(config), config, "config: ")
-
-   def escalate(self, res):
-      "Try to escalate authorization; return True if successful, else False."
-      auth = self.auth.escalate(self, res)
-      if (auth is None):
-         return False
-      else:
-         self.auth = auth
-         return True
-
-   def fatman_to_file(self, path, msg):
-      """GET the manifest for self.image and save it at path. This seems to
-         have four possible results:
-
-            1. HTTP 200, and body is a fat manifest: image exists and is
-               architecture-aware.
-
-            2. HTTP 200, but body is a skinny manifest: image exists but is
-               not architecture-aware.
-
-            3. HTTP 401/404: image does not exist or is unauthorized.
-
-            4. HTTP 429: rate limite exceeded.
-
-         This method raises Image_Unavailable_Error in case 3. The caller is
-         responsible for distinguishing cases 1 and 2."""
-      url = self._url_of("manifests", self.ref.version)
-      pw = Progress_Writer(path, msg)
-      # Including TYPES_MANIFEST avoids the server trying to convert its v2
-      # manifest to a v1 manifest, which currently fails for images
-      # Charliecloud pushes. The error in the test registry is “empty history
-      # when trying to create schema1 manifest”.
-      accept = "%s;q=0.5" % ",".join(  list(TYPES_INDEX.values())
-                                     + list(TYPES_MANIFEST.values()))
-      res = self.request("GET", url, out=pw, statuses={200, 401, 404, 429},
-                         headers={ "Accept" : accept })
-      pw.close()
-      if (res.status_code == 429):
-         if (self.auth.anon_p):
-            hint = "consider --auth"
-         else:
-            hint = None
-         FATAL("registry rate limit exceeded (HTTP 429)", hint)
-      elif (res.status_code != 200):
-         DEBUG(res.content)
-         raise Image_Unavailable_Error()
-
-   def layer_from_file(self, digest, path, note=""):
-      "Upload gzipped tarball layer at path, which must have hash digest."
-      # NOTE: We don't verify the digest b/c that means reading the whole file.
-      VERBOSE("layer tarball: %s" % path)
-      fp = path.open_("rb") # open file avoids reading it all into memory
-      self.blob_upload(digest, fp, note)
-      close_(fp)
-
-   def manifest_to_file(self, path, msg, digest=None):
-      """GET manifest for the image and save it at path. If digest is given,
-         use that to fetch the appropriate architecture; otherwise, fetch the
-         default manifest using the exising image reference."""
-      if (digest is None):
-         digest = self.ref.version
-      else:
-         digest = "sha256:" + digest
-      url = self._url_of("manifests", digest)
-      pw = Progress_Writer(path, msg)
-      accept = "%s;q=0.5" % ",".join(TYPES_MANIFEST.values())
-      res = self.request("GET", url, out=pw, statuses={200, 401, 404},
-                         headers={ "Accept" : accept })
-      pw.close()
-      if (res.status_code != 200):
-         DEBUG(res.content)
-         raise Image_Unavailable_Error()
-
-   def manifest_upload(self, manifest):
-      "Upload manifest (sequence of bytes)."
-      # Note: The manifest is *not* uploaded as a blob. We just do one PUT.
-      INFO("manifest: uploading")
-      url = self._url_of("manifests", self.ref.tag)
-      self.request("PUT", url, {201}, data=manifest,
-                   headers={ "Content-Type": TYPES_MANIFEST["docker2"] })
-
-   def request(self, method, url, statuses={200}, out=None, **kwargs):
-      """Request url using method and return the response object. If statuses
-         is given, it is set of acceptable response status codes, defaulting
-         to {200}; any other response is a fatal error. If out is given,
-         response content will be streamed to this Progress_Writer object and
-         must be non-zero length.
-
-         Use current session if there is one, or start a new one if not. If
-         authentication fails (or isn't initialized), then authenticate harder
-         and re-try the request."""
-      # Set up.
-      self.session_init_maybe()
-      VERBOSE("auth: %s" % self.auth)
-      if (out is not None):
-         kwargs["stream"] = True
-      # Make the request.
-      while True:
-         res = self.request_raw(method, url, statuses | {401}, **kwargs)
-         if (res.status_code != 401):
-            break
-         else:
-            VERBOSE("HTTP 401 unauthorized")
-            if (self.escalate(res)):   # success
-               VERBOSE("retrying with auth: %s" % self.auth)
-            elif (401 in statuses):    # caller can deal with it
-               break
-            else:
-               FATAL("unhandled authentication failure")
-      # Stream response if needed.
-      if (out is not None and res.status_code == 200):
-         try:
-            length = int(res.headers["Content-Length"])
-         except KeyError:
-            length = None
-         except ValueError:
-            FATAL("invalid Content-Length in response")
-         out.start(length)
-         for chunk in res.iter_content(HTTP_CHUNK_SIZE):
-            out.write(chunk)
-      # Done.
-      return res
-
-   def request_raw(self, method, url, statuses, auth=None, **kwargs):
-      """Request url using method. statuses is an iterable of acceptable
-         response status codes; any other response is a fatal error. Return
-         the requests.Response object.
-
-         Session must already exist. If auth arg given, use it; otherwise, use
-         object's stored authentication if initialized; otherwise, use no
-         authentication."""
-      VERBOSE("%s: %s" % (method, url))
-      if (auth is None):
-         auth = self.auth
-      try:
-         res = self.session.request(method, url, auth=auth, **kwargs)
-         VERBOSE("response status: %d" % res.status_code)
-         if (res.status_code not in statuses):
-            FATAL("%s failed; expected status %s but got %d: %s"
-                  % (method, statuses, res.status_code, res.reason))
-      except requests.exceptions.RequestException as x:
-         FATAL("%s failed: %s" % (method, x))
-      # Log some headers if needed.
-      for h in res.headers:
-         h = h.lower()
-         if (   "ratelimit" in h
-             or h == "www-authenticate"):
-            f = VERBOSE
-         else:
-            f = DEBUG
-         f("%s: %s" % (h, res.headers[h]))
-      return res
-
-   def session_init_maybe(self):
-      "Initialize session if it's not initialized; otherwise do nothing."
-      if (self.session is None):
-         VERBOSE("initializing session")
-         self.session = requests.Session()
-         self.session.verify = tls_verify
-
-
-class Storage:
-
-   """Source of truth for all paths within the storage directory. Do not
-      compute any such paths elsewhere!"""
-
-   __slots__ = ("lockfile_fp",
-                "root")
-
-   def __init__(self, storage_cli):
-      self.root = storage_cli
-      if (self.root is None):
-         self.root = self.root_env()
-      if (self.root is None):
-         self.root = self.root_default()
-      if (not self.root.is_absolute()):
-         self.root = os.getcwd() // self.root
-
-   @property
-   def build_cache(self):
-      return self.root // "bucache"
-
-   @property
-   def build_large(self):
-      return self.root // "bularge"
-
-   @property
-   def download_cache(self):
-      return self.root // "dlcache"
-
-   @property
-   def image_tmp(self):
-      return self.root // "imgtmp"
-
-   @property
-   def lockfile(self):
-      return self.root // "lock"
-
-   @property
-   def mount_point(self):
-      return self.root // "mnt"
-
-   @property
-   def unpack_base(self):
-      return self.root // "img"
-
-   @property
-   def upload_cache(self):
-      return self.root // "ulcache"
-
-   @property
-   def valid_p(self):
-      """Return True if storage present and seems valid, even if old, False
-         otherwise. This answers “is the storage directory real”, not “can
-         this storage directory be used”; it should return True for more or
-         less any Charliecloud storage directory we might feasibly come
-         across, even if it can't be upgraded. See also #1147."""
-      return (os.path.isdir(self.unpack_base) and
-              os.path.isdir(self.download_cache))
-
-   @property
-   def version_file(self):
-      return self.root // "version"
-
-   @staticmethod
-   def root_default():
-      # FIXME: Perhaps we should use getpass.getuser() instead of the $USER
-      # environment variable? It seems a lot more robust. But, (1) we'd have
-      # to match it in some scripts and (2) it makes the documentation less
-      # clear becase we have to explain the fallback behavior.
-      return Path("/var/tmp/%s.ch" % user())
-
-   @staticmethod
-   def root_env():
-      if ("CH_GROW_STORAGE" in os.environ):
-         # Avoid surprises if user still has $CH_GROW_STORAGE set (see #906).
-         FATAL("$CH_GROW_STORAGE no longer supported; use $CH_IMAGE_STORAGE")
-      if (not "CH_IMAGE_STORAGE" in os.environ):
-         return None
-      path = Path(os.environ["CH_IMAGE_STORAGE"])
-      if (not path.is_absolute()):
-         FATAL("$CH_IMAGE_STORAGE: not absolute path: %s" % path)
-      return path
-
-   def build_large_path(self, name):
-      return self.build_large // name
-
-   def init(self):
-      """Ensure the storage directory exists, contains all the appropriate
-         top-level directories & metadata, and is the appropriate version."""
-      # WARNING: This function contains multiple calls to self.lock(). The
-      # point is to lock as soon as we know the storage directory exists, and
-      # definitely before writing anything, to reduce the race conditions that
-      # surely exist. Ensure new code paths also call self.lock().
-      self.init_move_old()  # see issues #1160 and #1243
-      if (not os.path.isdir(self.root)):
-         op = "initializing"
-         v_found = None
-      else:
-         op = "upgrading"  # not used unless upgrading
-         if (not self.valid_p):
-            if (os.path.exists(self.root) and not self.root.listdir()):
-               hint = "let Charliecloud create %s; see FAQ" % self.root.name
-            else:
-               hint = None
-            FATAL("storage directory seems invalid: %s" % self.root, hint=hint)
-         v_found = self.version_read()
-      if (v_found == STORAGE_VERSION):
-         VERBOSE("found storage dir v%d: %s" % (STORAGE_VERSION, self.root))
-         self.lock()
-      elif (v_found in {None, 1, 2, 3}):  # initialize/upgrade
-         INFO("%s storage directory: v%d %s" % (op, STORAGE_VERSION, self.root))
-         self.root.mkdir_()
-         self.lock()
-         self.download_cache.mkdir_()
-         self.build_cache.mkdir_()
-         self.build_large.mkdir_()
-         self.unpack_base.mkdir_()
-         self.upload_cache.mkdir_()
-         for old in self.unpack_base.iterdir():
-            new = old.parent // str(old.name).replace(":", "+")
-            if (old != new):
-               if (new.exists()):
-                  FATAL("can't upgrade: already exists: %s" % new)
-               old.rename(new)
-         self.version_file.file_write("%d\n" % STORAGE_VERSION)
-      else:                         # can't upgrade
-         FATAL("incompatible storage directory v%d: %s" % (v_found, self.root),
-               'you can delete and re-initialize with "ch-image reset"')
-      self.validate_strict()
-
-   def init_move_old(self):
-      """If appropriate, move storage directory from old default path to new.
-         See issues #1160 and #1243."""
-      old = Storage(Path("/var/tmp") // user() // "ch-image")
-      moves = ( "dlcache", "img", "ulcache", "version" )
-      if (self.root != self.root_default()):
-         return  # do nothing silently unless using default storage dir
-      if (not os.path.exists(old.root)):
-         return  # do nothing silently unless old default storage dir exists
-      if (not old.valid_p):
-         WARNING("storage dir: invalid at old default, ignoring: %s" % old.root)
-         return
-      INFO("storage dir: valid at old default: %s" % old.root)
-      if (not os.path.exists(self.root)):
-         self.root.mkdir_()
-      elif (self.valid_p):
-         WARNING("storage dir: also valid at new default: %s" % self.root,
-                 hint="consider deleting the old one")
-         return
-      elif (not os.path.isdir(self.root)):
-         return  # new isn't a directory; init/upgrade code will error later
-      elif (any(os.path.exists(self.root // i) for i in moves)):
-         return  # new is broken; init/upgrade should error later
-      # Now we know (1) the old storage exists and is valid and (2) the new
-      # storage exists, is a directory, and contains none of the files we'd
-      # move. However, it *may* contain subdirectories other parts of
-      # Charliecloud care about, e.g. "mnt" for ch-run.
-      INFO("storage dir: moving to new default path: %s" % self.root)
-      for i in moves:
-         src = old.root // i
-         dst = self.root // i
-         if (os.path.exists(src)):
-            DEBUG("moving: %s -> %s" % (src, dst))
-            try:
-               shutil.move(src, dst)
-            except OSError as x:
-               FATAL("can't move: %s -> %s: %s"
-                     % (x.filename, x.filename2, x.strerror))
-      old.root.rmdir_()
-      if (not old.root.parent.listdir()):
-         WARNING("parent of old storage dir now empty: %s" % old.root.parent,
-                 hint="consider deleting it")
-
-   def lock(self):
-      """Lock the storage directory. Charliecloud does not at present support
-         concurrent use of ch-image(1) against the same storage directory."""
-      # File locking on Linux is a disaster [1, 2]. Currently, we use POSIX
-      # fcntl(2) locking, which has major pitfalls but should be fine for our
-      # use case. It apparently works on NFS [3] and does not require
-      # cleanup/stealing like a lock file would.
-      #
-      # [1]: https://apenwarr.ca/log/20101213
-      # [2]: http://0pointer.de/blog/projects/locking.html
-      # [3]: https://stackoverflow.com/a/22411531
-      if (not storage_lock):
-         return
-      self.lockfile_fp = self.lockfile.open_("w")
-      try:
-         fcntl.lockf(self.lockfile_fp, fcntl.LOCK_EX | fcntl.LOCK_NB)
-      except OSError as x:
-         if (x.errno in { errno.EACCES, errno.EAGAIN }):
-            FATAL("storage directory is already in use",
-                  "concurrent instances of ch-image cannot share the same storage directory")
-         else:
-            FATAL("can't lock storage directory: %s" % x.strerror)
-
-   def manifest_for_download(self, image_ref, digest):
-      if (digest is None):
-         digest = "skinny"
-      return (   self.download_cache
-              // ("%s%%%s.manifest.json" % (image_ref.for_path, digest)))
-
-   def fatman_for_download(self, image_ref):
-      return self.download_cache // ("%s.fat.json" % image_ref.for_path)
-
-   def reset(self):
-      if (self.valid_p):
-         self.root.rmtree()
-         self.init()  # largely for debugging
-      else:
-         FATAL("%s not a builder storage" % (self.root));
-
-   def unpack(self, image_ref):
-      return self.unpack_base // image_ref.for_path
-
-   def validate_strict(self):
-      """Validate storage directory structure; if something is wrong, exit
-         with an error message. This is a strict validation; the version must
-         be current, the structure of the directory must be current, and
-         nothing unexpected may be present. However, it is not comprehensive.
-         The main purpose is to check for bad upgrades and other programming
-         errors, not meddling."""
-      DEBUG("validating storage directory: %s" % self.root)
-      msg_prefix = "invalid storage directory"
-      # Check that all expected files exist, and no others. Note that we don't
-      # verify file *type*, assuming that kind of error is rare.
-      entries = self.root.listdir()
-      for entry in { i.name for i in (self.build_cache,
-                                      self.build_large,
-                                      self.download_cache,
-                                      self.unpack_base,
-                                      self.upload_cache,
-                                      self.version_file) }:
-         try:
-            entries.remove(entry)
-         except KeyError:
-            FATAL("%s: missing file or directory: %s" % (msg_prefix, entry))
-      entries -= { i.name for i in (self.lockfile, self.mount_point) }
-      if (len(entries) > 0):
-         FATAL("%s: extraneous file(s): %s"
-               % (msg_prefix, " ".join(sorted(entries))))
-      # check version
-      v_found = self.version_read()
-      if (v_found != STORAGE_VERSION):
-         FATAL("%s: version mismatch: %d expected, %d found"
-               % (msg_prefix, STORAGE_VERSION, v_found))
-      # check that no image directories have “:” in filename
-      assert isinstance(self.unpack_base, Path) # remove if test suite passes
-      imgs = self.unpack_base.listdir()
-      imgs_bad = set()
-      for img in imgs:
-         if (":" in img):  # bad char check b/c problem here is bad upgrade
-            FATAL("%s: storage directory broken: bad image dir name: %s"
-                  % (msg_prefix, img), BUG_REPORT_PLZ)
-
-   def version_read(self):
-      if (os.path.isfile(self.version_file)):
-         text = self.version_file.file_read_all() # WARNING: version_file might not
-                                                  # be Path
-         try:
-            return int(text)
-         except ValueError:
-            FATAL('malformed storage version: "%s"' % text)
-      else:
-         return 1
-
-
-class TarFile(tarfile.TarFile):
-
-   # This subclass augments tarfile.TarFile to add safety code. While the
-   # tarfile module docs [1] say “do not use this class [TarFile] directly”,
-   # they also say “[t]he tarfile.open() function is actually a shortcut” to
-   # class method TarFile.open(), and the source code recommends subclassing
-   # TarFile [2].
-   #
-   # It's here because the standard library class has problems with symlinks
-   # and replacing one file type with another; see issues #819 and #825 as
-   # well as multiple unfixed Python bugs [e.g. 3,4,5]. We work around this
-   # with manual deletions.
-   #
-   # [1]: https://docs.python.org/3/library/tarfile.html
-   # [2]: https://github.com/python/cpython/blob/2bcd0fe7a5d1a3c3dd99e7e067239a514a780402/Lib/tarfile.py#L2159
-   # [3]: https://bugs.python.org/issue35483
-   # [4]: https://bugs.python.org/issue19974
-   # [5]: https://bugs.python.org/issue23228
-
-   # Need new method name because add() is called recursively and we don't
-   # want those internal calls to get our special sauce.
-   def add_(self, name, **kwargs):
-      def filter_(ti):
-         assert (ti.name == "." or ti.name[:2] == "./")
-         if (ti.name.startswith("./.git") or ti.name == "./ch/git.pickle"):
-            DEBUG("omitting from push: %s" % ti.name)
-            return None
-         self.fix_member_uidgid(ti)
-         return ti
-      kwargs["filter"] = filter_
-      super().add(name, **kwargs)
-
-   def clobber(self, targetpath, regulars=False, symlinks=False, dirs=False):
-      assert (regulars or symlinks or dirs)
-      try:
-         st = os.lstat(targetpath)
-      except FileNotFoundError:
-         # We could move this except clause after all the stat.S_IS* calls,
-         # but that risks catching FileNotFoundError that came from somewhere
-         # other than lstat().
-         st = None
-      except OSError as x:
-         FATAL("can't lstat: %s" % targetpath, targetpath)
-      if (st is not None):
-         if (stat.S_ISREG(st.st_mode)):
-            if (regulars):
-               Path(targetpath).unlink_()
-         elif (stat.S_ISLNK(st.st_mode)):
-            if (symlinks):
-               Path(targetpath).unlink_()
-         elif (stat.S_ISDIR(st.st_mode)):
-            if (dirs):
-               Path(targetpath).rmtree()
-         else:
-            FATAL("invalid file type 0%o in previous layer; see inode(7): %s"
-                  % (stat.S_IFMT(st.st_mode), targetpath))
-
-   @staticmethod
-   def fix_link_target(ti, tb):
-      """Deal with link (symbolic or hard) weirdness or breakage. If it can be
-         fixed, fix it; if not, abort the program."""
-      src = Path(ti.name)
-      tgt = Path(ti.linkname)
-      fix_ct = 0
-      # Empty target not allowed; have to check string b/c "" -> Path(".").
-      if (len(ti.linkname) == 0):
-         FATAL("rejecting link with empty target: %s: %s" % (tb, ti.name))
-      # Fix absolute link targets.
-      if (tgt.is_absolute()):
-         if (ti.issym()):
-            # Change symlinks to relative for correct interpretation inside or
-            # outside the container.
-            kind = "symlink"
-            new = (  Path(*(("..",) * (len(src.parts) - 1)))
-                   // Path(*(tgt.parts[1:])))
-         elif (ti.islnk()):
-            # Hard links refer to tar member paths; just strip leading slash.
-            kind = "hard link"
-            new = tgt.relative_to("/")
-         else:
-            assert False, "not a link"
-         DEBUG("absolute %s: %s -> %s: changing target to: %s"
-               % (kind, src, tgt, new))
-         tgt = new
-         fix_ct = 1
-      # Reject links that climb out of image (FIXME: repair instead).
-      if (".." in os.path.normpath(src // tgt).split("/")):
-         FATAL("rejecting too many up-levels: %s: %s -> %s" % (tb, src, tgt))
-      # Done.
-      ti.linkname = str(tgt)
-      return fix_ct
-
-   @staticmethod
-   def fix_member_uidgid(ti):
-      assert (ti.name[0] != "/")  # absolute paths unsafe but shouldn't happen
-      if (not (ti.isfile() or ti.isdir() or ti.issym() or ti.islnk())):
-         FATAL("invalid file type: %s" % ti.name)
-      ti.uid = 0
-      ti.uname = "root"
-      ti.gid = 0
-      ti.gname = "root"
-      if (ti.mode & stat.S_ISUID):
-         VERBOSE("stripping unsafe setuid bit: %s" % ti.name)
-         ti.mode &= ~stat.S_ISUID
-      if (ti.mode & stat.S_ISGID):
-         VERBOSE("stripping unsafe setgid bit: %s" % ti.name)
-         ti.mode &= ~stat.S_ISGID
-
-   def makedir(self, tarinfo, targetpath):
-      # Note: This gets called a lot, e.g. once for each component in the path
-      # of the member being extracted.
-      TRACE("makedir: %s" % targetpath)
-      self.clobber(targetpath, regulars=True, symlinks=True)
-      super().makedir(tarinfo, targetpath)
-
-   def makefile(self, tarinfo, targetpath):
-      TRACE("makefile: %s" % targetpath)
-      self.clobber(targetpath, symlinks=True, dirs=True)
-      super().makefile(tarinfo, targetpath)
-
-   def makelink(self, tarinfo, targetpath):
-      TRACE("makelink: %s -> %s" % (targetpath, tarinfo.linkname))
-      self.clobber(targetpath, regulars=True, symlinks=True, dirs=True)
-      super().makelink(tarinfo, targetpath)
 
 class Timer:
